fix: item 가격, 보유 sands, 총 로그인 일수 데이터타입 변경 (Long -> int)

+++ dolharubang/src/main/java/com/dolharubang/controller/MemberController.java
@@ -77,7 +77,7 @@ public class MemberController {
 
     @Operation(summary = "회원 모래알 조회하기", description = "모래알을 조회한다.")
     @GetMapping("/sands/{memberId}")
-    public Long getSands(@PathVariable Long memberId) {
+    public int getSands(@PathVariable Long memberId) {
         return memberService.getSands(memberId);
     }
 }
